+++ crates/matrix-sdk-crypto/src/session_manager/group_sessions/share_strategy.rs
@@ -120,6 +120,7 @@ pub(crate) async fn collect_session_recipients(
     let users: BTreeSet<&UserId> = users.collect();
     let mut devices: BTreeMap<OwnedUserId, Vec<DeviceData>> = Default::default();
     let mut withheld_devices: Vec<(DeviceData, WithheldCode)> = Default::default();
+    let mut verified_users_with_new_identities: Vec<OwnedUserId> = Default::default();
 
     trace!(?users, ?settings, "Calculating group session recipients");
 
@@ -145,6 +146,8 @@ pub(crate) async fn collect_session_recipients(
     // This is calculated in the following code and stored in this variable.
     let mut should_rotate = user_left || visibility_changed || algorithm_changed;
 
+    let own_identity = store.get_user_identity(store.user_id()).await?.and_then(|i| i.into_own());
+
     // Get the recipient and withheld devices, based on the collection strategy.
     match settings.sharing_strategy {
         CollectStrategy::DeviceBasedStrategy {
@@ -153,10 +156,6 @@ pub(crate) async fn collect_session_recipients(
         } => {
             let mut unsigned_devices_of_verified_users: BTreeMap<OwnedUserId, Vec<OwnedDeviceId>> =
                 Default::default();
-            let mut verified_users_with_new_identities: Vec<OwnedUserId> = Default::default();
-
-            let own_identity =
-                store.get_user_identity(store.user_id()).await?.and_then(|i| i.into_own());
 
             for user_id in users {
                 trace!("Considering recipient devices for user {}", user_id);
@@ -233,23 +232,38 @@ pub(crate) async fn collect_session_recipients(
                     ),
                 ));
             }
-
-            // Alternatively, we may have encountered previously-verified users who have
-            // changed their identities. We bail out for that, too.
-            if !verified_users_with_new_identities.is_empty() {
-                return Err(OlmError::SessionRecipientCollectionError(
-                    SessionRecipientCollectionError::VerifiedUserChangedIdentity(
-                        verified_users_with_new_identities,
-                    ),
-                ));
-            }
         }
         CollectStrategy::IdentityBasedStrategy => {
+            // We require our own cross-signing to be properly set up for the
+            // identity-based strategy, so return an error if it isn't.
+            match &own_identity {
+                None => {
+                    return Err(OlmError::SessionRecipientCollectionError(
+                        SessionRecipientCollectionError::CrossSigningNotSetup,
+                    ))
+                }
+                Some(identity) if !identity.is_verified() => {
+                    return Err(OlmError::SessionRecipientCollectionError(
+                        SessionRecipientCollectionError::SendingFromUnverifiedDevice,
+                    ))
+                }
+                Some(_) => (),
+            }
+
             for user_id in users {
                 trace!("Considering recipient devices for user {}", user_id);
                 let user_devices = store.get_device_data_for_user_filtered(user_id).await?;
 
                 let device_owner_identity = store.get_user_identity(user_id).await?;
+
+                if has_identity_verification_violation(
+                    own_identity.as_ref(),
+                    device_owner_identity.as_ref(),
+                ) {
+                    verified_users_with_new_identities.push(user_id.to_owned());
+                    // No point considering the individual devices of this user.
+                    continue;
+                }
 
                 let recipient_devices = split_recipients_withhelds_for_user_based_on_identity(
                     user_devices,
@@ -275,6 +289,16 @@ pub(crate) async fn collect_session_recipients(
                 withheld_devices.extend(recipient_devices.denied_devices_with_code);
             }
         }
+    }
+
+    // We may have encountered previously-verified users who have changed their
+    // identities. If so, we bail out with an error.
+    if !verified_users_with_new_identities.is_empty() {
+        return Err(OlmError::SessionRecipientCollectionError(
+            SessionRecipientCollectionError::VerifiedUserChangedIdentity(
+                verified_users_with_new_identities,
+            ),
+        ));
     }
 
     if should_rotate {
@@ -491,10 +515,14 @@ fn is_user_verified(
 mod tests {
     use std::{collections::BTreeMap, iter, sync::Arc};
 
+    use assert_matches::assert_matches;
     use assert_matches2::assert_let;
     use matrix_sdk_test::{
         async_test, test_json,
-        test_json::keys_query_sets::{KeyDistributionTestData, PreviouslyVerifiedTestData},
+        test_json::keys_query_sets::{
+            IdentityChangeDataSet, KeyDistributionTestData, MaloIdentityChangeDataSet,
+            PreviouslyVerifiedTestData,
+        },
     };
     use ruma::{
         device_id, events::room::history_visibility::HistoryVisibility, room_id, TransactionId,
@@ -1120,6 +1148,249 @@ mod tests {
             .expect("This dave device should receive a withheld code");
 
         assert_eq!(code, &WithheldCode::Unauthorised);
+    }
+
+    /// Test key sharing with the identity-based strategy with different
+    /// states of our own verification.
+    #[async_test]
+    async fn test_share_identity_strategy_no_cross_signing() {
+        // Starting off, we have not yet set up our own cross-signing, so
+        // sharing with the identity-based strategy should fail.
+        let machine: OlmMachine = OlmMachine::new(
+            KeyDistributionTestData::me_id(),
+            KeyDistributionTestData::me_device_id(),
+        )
+        .await;
+
+        let keys_query = KeyDistributionTestData::dan_keys_query_response();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+
+        let fake_room_id = room_id!("!roomid:localhost");
+
+        let encryption_settings = EncryptionSettings {
+            sharing_strategy: CollectStrategy::new_identity_based(),
+            ..Default::default()
+        };
+
+        let request_result = machine
+            .share_room_key(
+                fake_room_id,
+                iter::once(KeyDistributionTestData::dan_id()),
+                encryption_settings.clone(),
+            )
+            .await;
+
+        assert_matches!(
+            request_result,
+            Err(OlmError::SessionRecipientCollectionError(
+                SessionRecipientCollectionError::CrossSigningNotSetup
+            ))
+        );
+
+        // We now get our public cross-signing keys, but we don't trust them
+        // yet.  In this case, sharing the keys should still fail since our own
+        // device is still unverified.
+        let keys_query = KeyDistributionTestData::me_keys_query_response();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+
+        let request_result = machine
+            .share_room_key(
+                fake_room_id,
+                iter::once(KeyDistributionTestData::dan_id()),
+                encryption_settings.clone(),
+            )
+            .await;
+
+        assert_matches!(
+            request_result,
+            Err(OlmError::SessionRecipientCollectionError(
+                SessionRecipientCollectionError::SendingFromUnverifiedDevice
+            ))
+        );
+
+        // Finally, after we trust our own cross-signing keys, key sharing
+        // should succeed.
+        machine
+            .import_cross_signing_keys(CrossSigningKeyExport {
+                master_key: KeyDistributionTestData::MASTER_KEY_PRIVATE_EXPORT.to_owned().into(),
+                self_signing_key: KeyDistributionTestData::SELF_SIGNING_KEY_PRIVATE_EXPORT
+                    .to_owned()
+                    .into(),
+                user_signing_key: KeyDistributionTestData::USER_SIGNING_KEY_PRIVATE_EXPORT
+                    .to_owned()
+                    .into(),
+            })
+            .await
+            .unwrap();
+
+        let requests = machine
+            .share_room_key(
+                fake_room_id,
+                iter::once(KeyDistributionTestData::dan_id()),
+                encryption_settings.clone(),
+            )
+            .await
+            .unwrap();
+
+        // Dan has two devices, but only one is cross-signed, so there should
+        // only be one key share.
+        assert_eq!(requests.len(), 1);
+    }
+
+    /// Test that identity-based key sharing gives an error when a verified
+    /// user changes their identity, and that the key can be shared when the
+    /// identity change is resolved.
+    #[async_test]
+    async fn test_share_identity_strategy_report_verification_violation() {
+        let machine: OlmMachine = OlmMachine::new(
+            KeyDistributionTestData::me_id(),
+            KeyDistributionTestData::me_device_id(),
+        )
+        .await;
+
+        machine.bootstrap_cross_signing(false).await.unwrap();
+
+        // We will try sending a key to two different users.
+        let user1 = IdentityChangeDataSet::user_id();
+        let user2 = MaloIdentityChangeDataSet::user_id();
+
+        // We first get both users' initial device and identity keys.
+        let keys_query = IdentityChangeDataSet::key_query_with_identity_a();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+
+        let keys_query = MaloIdentityChangeDataSet::initial_key_query();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+
+        // And then we get both user' changed identity keys.  We simulate a
+        // verification violation by marking both users as having been
+        // previously verified, in which case the key sharing should fail.
+        let keys_query = IdentityChangeDataSet::key_query_with_identity_b();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+        machine
+            .get_identity(user1, None)
+            .await
+            .unwrap()
+            .unwrap()
+            .other()
+            .unwrap()
+            .mark_as_previously_verified()
+            .await
+            .unwrap();
+
+        let keys_query = MaloIdentityChangeDataSet::updated_key_query();
+        machine.mark_request_as_sent(&TransactionId::new(), &keys_query).await.unwrap();
+        machine
+            .get_identity(user2, None)
+            .await
+            .unwrap()
+            .unwrap()
+            .other()
+            .unwrap()
+            .mark_as_previously_verified()
+            .await
+            .unwrap();
+
+        let fake_room_id = room_id!("!roomid:localhost");
+
+        // We share the key using the identity-based strategy.
+        let encryption_settings = EncryptionSettings {
+            sharing_strategy: CollectStrategy::new_identity_based(),
+            ..Default::default()
+        };
+
+        let request_result = machine
+            .share_room_key(
+                fake_room_id,
+                vec![user1, user2].into_iter(),
+                encryption_settings.clone(),
+            )
+            .await;
+
+        // The key share should fail with an error indicating that recipients
+        // were previously verified.
+        assert_let!(
+            Err(OlmError::SessionRecipientCollectionError(
+                SessionRecipientCollectionError::VerifiedUserChangedIdentity(affected_users)
+            )) = request_result
+        );
+        // Both our recipients should be in `affected_users`.
+        assert_eq!(2, affected_users.len());
+
+        // We resolve this for user1 by withdrawing their verification.
+        machine
+            .get_identity(user1, None)
+            .await
+            .unwrap()
+            .unwrap()
+            .withdraw_verification()
+            .await
+            .unwrap();
+
+        // We resolve this for user2 by re-verifying.
+        let verification_request = machine
+            .get_identity(user2, None)
+            .await
+            .unwrap()
+            .unwrap()
+            .other()
+            .unwrap()
+            .verify()
+            .await
+            .unwrap();
+        let raw_extracted =
+            verification_request.signed_keys.get(user2).unwrap().iter().next().unwrap().1.get();
+        let signed_key: crate::types::CrossSigningKey =
+            serde_json::from_str(raw_extracted).unwrap();
+        let new_signatures = signed_key.signatures.get(KeyDistributionTestData::me_id()).unwrap();
+        let mut master_key = machine
+            .get_identity(user2, None)
+            .await
+            .unwrap()
+            .unwrap()
+            .other()
+            .unwrap()
+            .master_key
+            .as_ref()
+            .clone();
+
+        for (key_id, signature) in new_signatures.iter() {
+            master_key.as_mut().signatures.add_signature(
+                KeyDistributionTestData::me_id().to_owned(),
+                key_id.to_owned(),
+                signature.as_ref().unwrap().ed25519().unwrap(),
+            );
+        }
+        let json = serde_json::json!({
+            "device_keys": {},
+            "failures": {},
+            "master_keys": {
+                user2: master_key,
+            },
+            "user_signing_keys": {},
+            "self_signing_keys": MaloIdentityChangeDataSet::updated_key_query().self_signing_keys,
+        }
+        );
+
+        let kq_response = matrix_sdk_test::ruma_response_from_json(&json);
+        machine
+            .mark_request_as_sent(
+                &TransactionId::new(),
+                crate::IncomingResponse::KeysQuery(&kq_response),
+            )
+            .await
+            .unwrap();
+
+        assert!(machine.get_identity(user2, None).await.unwrap().unwrap().is_verified());
+
+        // And now the key share should succeed.
+        machine
+            .share_room_key(
+                fake_room_id,
+                vec![user1, user2].into_iter(),
+                encryption_settings.clone(),
+            )
+            .await
+            .unwrap();
     }
 
     #[async_test]