+++ testing/matrix-sdk-test/src/test_json/keys_query_sets.rs
@@ -1263,3 +1263,151 @@ impl PreviouslyVerifiedTestData {
         ruma_response_from_json(&data)
     }
 }
+
+/// A set of keys query to test identity changes,
+/// For user @malo, that performed an identity change with the same device.
+pub struct MaloIdentityChangeDataSet {}
+
+#[allow(dead_code)]
+impl MaloIdentityChangeDataSet {
+    pub fn user_id() -> &'static UserId {
+        user_id!("@malo:localhost")
+    }
+
+    pub fn device_id() -> &'static DeviceId {
+        device_id!("NZFSPBRLDO")
+    }
+
+    /// @malo's keys before their identity change
+    pub fn initial_key_query() -> KeyQueryResponse {
+        let data = json!({
+            "device_keys": {
+                "@malo:localhost": {
+                    "NZFSPBRLDO": {
+                        "algorithms": [
+                            "m.olm.v1.curve25519-aes-sha2",
+                            "m.megolm.v1.aes-sha2"
+                        ],
+                        "device_id": "NZFSPBRLDO",
+                        "keys": {
+                            "curve25519:NZFSPBRLDO": "L3jdbw42+9i+K7LPjAY+kmqG9nr2n/U0ow8hEbLCoCs",
+                            "ed25519:NZFSPBRLDO": "VDJt3xI4SzrgQkuE3sEIauluaXawx3wWoWOynPI8Zko"
+                        },
+                        "signatures": {
+                            "@malo:localhost": {
+                                "ed25519:NZFSPBRLDO": "lmtbdrJ5xBweo677Fg2qrSHsRi4R3x2WNlvSNJY6Zbg0R5lJS9syN2HZw/irL9PA644GYm4QM/t+DX0grnn+BQ",
+                                "ed25519:+wbxNfSuDrch1jKuydQmEf4qlA4u4NgwqNXNuLVwug8": "Ql1fq+SvVDx+8mjNMzSaR0hBCEkdPirbs2+BK0gwsIH1zkuMADnBoNWP7LJiKo/EO9gnpiCzyQQgI4e9pIVPDA"
+                            }
+                        },
+                        "user_id": "@malo:localhost",
+                        "unsigned": {}
+                    }
+                }
+            },
+            "failures": {},
+            "master_keys": {
+                "@malo:localhost": {
+                    "keys": {
+                        "ed25519:WBxliSP29guYr4ux0MW6otRe3V/wOLXXElpOcOmpdlE": "WBxliSP29guYr4ux0MW6otRe3V/wOLXXElpOcOmpdlE"
+                    },
+                    "signatures": {
+                        "@malo:localhost": {
+                            "ed25519:NZFSPBRLDO": "crJcXqFpEHRM8KNUw419XrVFaHoM8/kV4ebgpuuIiD9wfX0AhHE2iGRGpKzsrVCqne9k181/uN0sgDMpK2y4Aw",
+                            "ed25519:WBxliSP29guYr4ux0MW6otRe3V/wOLXXElpOcOmpdlE": "/xwFF5AC3GhkpvJ449Srh8kNQS6CXAxQMmBpQvPEHx5BHPXJ08u2ZDd1EPYY4zk4QsePk+tEYu8gDnB0bggHCA"
+                        }
+                    },
+                    "usage": [
+                        "master"
+                    ],
+                    "user_id": "@malo:localhost"
+                }
+            },
+            "self_signing_keys": {
+                "@malo:localhost": {
+                    "keys": {
+                        "ed25519:+wbxNfSuDrch1jKuydQmEf4qlA4u4NgwqNXNuLVwug8": "+wbxNfSuDrch1jKuydQmEf4qlA4u4NgwqNXNuLVwug8"
+                    },
+                    "signatures": {
+                        "@malo:localhost": {
+                            "ed25519:WBxliSP29guYr4ux0MW6otRe3V/wOLXXElpOcOmpdlE": "sSGQ6ny6aXtIvgKPGOYJzcmnNDSkbaJFVRe9wekOry7EaiWf2l28MkGTUBt4cPoRiMkNjuRBupNEARqHF72sAQ"
+                        }
+                    },
+                    "usage": [
+                        "self_signing"
+                    ],
+                    "user_id": "@malo:localhost"
+                }
+            },
+            "user_signing_keys": {},
+        });
+
+        ruma_response_from_json(&data)
+    }
+
+    /// @malo's keys after their identity change
+    pub fn updated_key_query() -> KeyQueryResponse {
+        let data = json!({
+            "device_keys": {
+                "@malo:localhost": {
+                    "NZFSPBRLDO": {
+                        "algorithms": [
+                            "m.olm.v1.curve25519-aes-sha2",
+                            "m.megolm.v1.aes-sha2"
+                        ],
+                        "device_id": "NZFSPBRLDO",
+                        "keys": {
+                            "curve25519:NZFSPBRLDO": "L3jdbw42+9i+K7LPjAY+kmqG9nr2n/U0ow8hEbLCoCs",
+                            "ed25519:NZFSPBRLDO": "VDJt3xI4SzrgQkuE3sEIauluaXawx3wWoWOynPI8Zko"
+                        },
+                        "signatures": {
+                            "@malo:localhost": {
+                                "ed25519:NZFSPBRLDO": "lmtbdrJ5xBweo677Fg2qrSHsRi4R3x2WNlvSNJY6Zbg0R5lJS9syN2HZw/irL9PA644GYm4QM/t+DX0grnn+BQ",
+                                "ed25519:+wbxNfSuDrch1jKuydQmEf4qlA4u4NgwqNXNuLVwug8": "Ql1fq+SvVDx+8mjNMzSaR0hBCEkdPirbs2+BK0gwsIH1zkuMADnBoNWP7LJiKo/EO9gnpiCzyQQgI4e9pIVPDA",
+                                "ed25519:8my6+zgnzEP0ZqmQFyvscJh7isHlf8lxBmHg+fzdJkE": "OvqDE7C2mrHxjwNyMIEz+m/AO6I6lM5HoPYY2bvLjrJJDOF5sJOtw4JoYiCWyt90ZIWsbEqmfbazrblLD50tCg"
+                            }
+                        },
+                        "user_id": "@malo:localhost",
+                        "unsigned": {}
+                    }
+                }
+            },
+            "failures": {},
+            "master_keys": {
+                "@malo:localhost": {
+                    "keys": {
+                        "ed25519:dv2Mk7bFlRtP/0oSZpB01Ouc5frCXKfG8Bn9YrFxbxU": "dv2Mk7bFlRtP/0oSZpB01Ouc5frCXKfG8Bn9YrFxbxU"
+                    },
+                    "signatures": {
+                        "@malo:localhost": {
+                            "ed25519:NZFSPBRLDO": "2Ye96l4srBSWskNQszuMpea1r97rFoUyfNqegvu/hGeP47w0OVvqYuNtZRNwqb7TMS7aPEn6l9lhWEk7v06wCg",
+                            "ed25519:dv2Mk7bFlRtP/0oSZpB01Ouc5frCXKfG8Bn9YrFxbxU": "btkxAJpJeVtc9wgBmeHUI9QDpojd6ddLxK11E3403KoTQtP6Mnr5GsVdQr1HJToG7PG4k4eEZGWxVZr1GPndAA"
+                        }
+                    },
+                    "usage": [
+                        "master"
+                    ],
+                    "user_id": "@malo:localhost"
+                }
+            },
+            "self_signing_keys": {
+                "@malo:localhost": {
+                    "keys": {
+                        "ed25519:8my6+zgnzEP0ZqmQFyvscJh7isHlf8lxBmHg+fzdJkE": "8my6+zgnzEP0ZqmQFyvscJh7isHlf8lxBmHg+fzdJkE"
+                    },
+                    "signatures": {
+                        "@malo:localhost": {
+                            "ed25519:dv2Mk7bFlRtP/0oSZpB01Ouc5frCXKfG8Bn9YrFxbxU": "KJt0y1p8v8RGLGk2wUyCMbX1irXJqup/mdRuG/cxJxs24BZhDMyIzyGrGXnWq2gx3I4fKIMtFPi/ecxf92ePAQ"
+                        }
+                    },
+                    "usage": [
+                        "self_signing"
+                    ],
+                    "user_id": "@malo:localhost"
+                }
+            },
+            "user_signing_keys": {}
+        });
+
+        ruma_response_from_json(&data)
+    }
+}